requirements.txt
python-snap7==1.1
pyodbc==5.2.0
python-dotenv==1.0.1

pandas==2.3.2
numpy==2.2.6
scikit-learn==1.7.2
lightgbm==4.6.0
xgboost==3.0.5

onnx==1.19.0
onnxruntime==1.22.1

fastapi==0.116.1
uvicorn==0.35.0

streamlit==1.49.1
plotly==6.3.0
altair==5.5.0

skl2onnx==1.19.1
onnx==1.19.0
onnxruntime==1.22.1

pydantic==2.9.2
python-multipart==0.0.9
loguru==0.7.2
slowapi==0.1.9

